+++ src/main/java/seedu/planitarium/money/Expenditure.java
@@ -1,4 +1,4 @@
-package seedu.planitarium.expenditure;
+package seedu.planitarium.money;
 
 
 public class Expenditure {
