+++ src/main/java/seedu/planitarium/money/MoneyList.java
@@ -1,4 +1,9 @@
 package seedu.planitarium.money;
 
+import seedu.planitarium.ProjectLogger;
+
+import java.util.ArrayList;
+import java.util.logging.Level;
+
 public abstract class MoneyList {
 }
