+++ src/main/java/seedu/planitarium/person/Person.java
@@ -1,7 +1,7 @@
 package seedu.planitarium.person;
 
-import seedu.planitarium.income.IncomeList;
-import seedu.planitarium.expenditure.ExpenditureList;
+import seedu.planitarium.money.IncomeList;
+import seedu.planitarium.money.ExpenditureList;
 
 public class Person {
     protected String name;
