+++ src/test/java/seedu/planitarium/expenditure/ExpenditureListTest.java
@@ -2,6 +2,8 @@ package seedu.planitarium.expenditure;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import seedu.planitarium.money.ExpenditureList;
+
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNull;
 import static org.junit.jupiter.api.Assertions.fail;
