+++ src/test/java/seedu/planitarium/expenditure/ExpenditureTest.java
@@ -2,6 +2,7 @@ package seedu.planitarium.expenditure;
 
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
+import seedu.planitarium.money.Expenditure;
 
 import static org.junit.jupiter.api.Assertions.assertEquals;
 import static org.junit.jupiter.api.Assertions.assertNull;
